% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/have_gaps.R
\name{gaps_count}
\alias{gaps_count}
\title{Numărul de ferestre dintr-o matrice-orar dată}
\usage{
gaps_count(OZ)
}
\arguments{
\item{OZ}{Matrice-orar}
}
\value{
Numărul de ferestre, considerând cuplajele și profesorii externi
}
\description{
Aplicație imediată a funcției 'have_gaps()'; servește de exemplu, pentru
a afla câte ferestre sunt pe matricea-orar inițială (pe care ar urma să se
aplice 'recast()', pentru a reduce numărul de ferestre).
}
