% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/have_gaps.R
\name{have_gaps}
\alias{have_gaps}
\title{Evidențiază orarele profesorilor care au ferestre}
\usage{
have_gaps(OZ)
}
\arguments{
\item{OZ}{Matrice-orar; liniile sunt numite după profesori/cuplaje; pe
fiecare linie sunt înregistrate pe ore clasele la care intră profesorul
(sau '-' dacă este liber); fiecare clasă apare o singură dată, pe
fiecare coloană de rang cel mult egal cu numărul de ore/zi al clasei.}
}
\value{
Subsetul liniilor cu ferestre; pentru cazul ferestrelor "ascunse"
(induse de angajarea în cuplaje), este anexat șablonul tuturor orelor
în care este implicat profesorul respectiv.
}
\description{
Se șablonează orarele individuale, înlocuind clasele cu câte un asterisc;
fereastră înseamnă atunci orice '-' aflat între două asteriscuri; iar
pentru cei angajați în cuplaje, se cumulează șabloanele membrilor.
}
\examples{
HG <- have_gaps(MOZ) 
}
