% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitgaps-package.R
\docType{package}
\name{refitgaps-package}
\alias{refitgaps}
\alias{refitgaps-package}
\title{refitgaps: Reduce the Number of Holes in the School Timetable}
\description{
Reallocating the respective lessons by hours (respecting the constraints induced by the existence of coupled lessons) so that the total number of gaps is as small as possible.
}
\author{
\strong{Maintainer}: Vlad Bazon \email{vlad.bazon@gmail.com}

}
\keyword{internal}
