% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TPL}
\alias{TPL}
\title{Tupled lessons for 'MOZ'}
\format{
\describe{
\item{prof}{Vectorul profesorilor sau cuplajelor din tuplaj}
\item{cls}{Vectorul claselor din tuplaj, de aceeași lungime cu 'prof';
pentru fiecare tuplaj asocierea 'prof/cls' (în cadrul matricei-orar)
respectă ordinea profesorilor și claselor din acel tuplaj}
\item{ora}{Ora alocată în matricea-orar lecțiilor din tuplaj}
}
}
\usage{
TPL
}
\description{
6 tuplaje asociate matricei-orar 'MOZ'.
Lecțiile dintr-un același tuplaj au fost alocate într-o aceeași oră.
}
\keyword{datasets}
