% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refitgaps-package.R
\docType{package}
\name{refitgaps-package}
\alias{refitgaps}
\alias{refitgaps-package}
\title{refitgaps: Reduce the Number of Holes in the School Timetable}
\description{
If a teacher was assigned, say, hours 1, 2, 4, 6 for his lessons on the current day, then he got two "gaps" (or "holes"), in hours 3 and 5. If the total number of holes in the existing daily schedule is too large (exceeding, say, 6% of the total lessons), then the problem arises of reallocating the respective lessons by hours (respecting the constraints induced by the existence of coupled lessons) so that the total number of holes is as small as possible.
}
\author{
\strong{Maintainer}: Vlad Bazon \email{vlad.bazon@gmail.com}

}
\keyword{internal}
